requirements.txt
certifi==2018.8.24
chardet==3.0.4
Click==7.0
cssselect==1.0.3
Flask==1.0.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
lxml==4.2.5
MarkupSafe==1.0
PyYAML==3.13
readability-lxml==0.7
requests==2.19.1
urllib3==1.23
Werkzeug==0.14.1
